chore: move widget to docs layout
